requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.9.1
    # via django
billiard==4.2.1
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
django==5.2.6
    # via
    #   -r requirements.in
    #   django-tenants
django-environ==0.12.0
    # via -r requirements.in
django-tenants==3.9.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   gunicorn
    #   kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-json-logger==3.3.0
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django
    #   kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.9.0
    # via -r requirements.in
